requirements.txt
django>=5.0
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary